+++ docs/release-notes.md
@@ -15,6 +15,7 @@ Minor changes:
 - Add release notes to documentation
 - Fix default dependency ordering on all `checkin` services
 - Fix failure setting SSH keys on IBM Cloud if none are provided
+- Don't ignore network interfaces that appear during DHCP option lookup retry
 
 Packaging changes: